requirements-vercel.txt
fastapi
groq
llama-index-core
llama-index-readers-file
python-dotenv
qdrant-client
sentence-transformers
requests
secure-smtplib
pandas
openpyxl
plotly
python-multipart
pyjwt
